refactor: refactor AddUserAddressComponent
